[FEAT]: PresentationModel과 Mapper 추가
