+++ Projects/Presentation/Challenge/Implementations/Member/Feed/PresentationModel/FeedPresentatoinModelMapper.swift
@@ -1,0 +1,79 @@
+//
+//  FeedPresentatoinModelMapper.swift
+//  ChallengeImpl
+//
+//  Created by jung on 2/26/25.
+//  Copyright © 2025 com.photi. All rights reserved.
+//
+
+import Foundation
+import Core
+import Entity
+
+struct FeedPresentatoinModelMapper {
+  func mapToFeedPresentationModels(_ feeds: [Feed]) -> [FeedPresentationModel] {
+    return feeds.map { feed in
+      return .init(
+        id: feed.id,
+        imageURL: feed.imageURL,
+        userName: feed.author,
+        updateTime: mapToUpdateTimeString(feed.updateTime),
+        updateGroup: mapToUpdateGroup(feed.updateTime),
+        isLike: feed.isLike
+      )
+    }
+  }
+  
+  func mapToAuthorPresentaionModel(author: String, url: URL?) -> AuthorPresentationModel {
+    return .init(name: author, imageURL: url)
+  }
+}
+
+extension FeedPresentatoinModelMapper {
+  func mapToUpdateTimeString(_ date: Date) -> String {
+    let date = date.convertTimezone(from: .kst)
+    let current = Date()
+  
+    guard current.year == date.year else {
+      return "\(abs(current.year - date.year))년 전"
+    }
+  
+    guard current.month == date.month else {
+      return "\(abs(current.month - date.month))년 전"
+    }
+  
+    guard current.day == date.day else {
+      return "\(abs(current.day - date.day))일 전"
+    }
+  
+    guard current.hour == date.hour else {
+      return "\(abs(current.hour - date.hour))시간 전"
+    }
+  
+    guard current.minute != date.minute else {
+      return "방금"
+    }
+  
+    let temp = abs(current.minute - date.minute)
+    switch temp {
+      case 0...10: return "\(temp)분 전"
+      default: return "\(temp / 10)분 전"
+    }
+  }
+  
+  func mapToUpdateGroup(_ date: Date) -> String {
+    let date = date.convertTimezone(from: .kst)
+    let current = Date()
+  
+    guard current.year == date.year else {
+      return "\(abs(current.year - date.year))년 전"
+    }
+  
+    guard current.month == date.month else {
+      return "\(abs(current.month - date.month))년 전"
+    }
+  
+    let temp = abs(current.day - date.day)
+    return temp == 0 ? "오늘" : "\(temp)일 전"
+  }
+}
